fix up warnings
